+++ packages/gallery/src/components/item/itemView.js
@@ -241,8 +241,9 @@ class ItemView extends GalleryComponent {
 
   shouldHover() { //see if this could be decided in the preset
     const { styleParams } = this.props;
-    const { alwaysShowHover, previewHover, hoveringBehaviour } = styleParams;
+    const { alwaysShowHover, previewHover, hoveringBehaviour, overlayAnimation } = styleParams;
     const { NEVER_SHOW, APPEARS } = GALLERY_CONSTS.infoBehaviourOnHover;
+    const { NO_EFFECT } = GALLERY_CONSTS.overlayAnimations;
 
     if (hoveringBehaviour === NEVER_SHOW) {
       return false;
@@ -250,7 +251,12 @@ class ItemView extends GalleryComponent {
       return true;
     } else if (isEditMode() && previewHover) {
       return true;
-    } else if (!this.state.itemWasHovered && hoveringBehaviour === APPEARS) {
+    } else if (
+      hoveringBehaviour === APPEARS &&
+      overlayAnimation === NO_EFFECT &&
+      !this.state.itemWasHovered) {
+      //when there is no overlayAnimation, we want to render the itemHover only on first hover and on (and not before)
+      //when there is a specific overlayAnimation, to support the animation we should render the itemHover before any hover activity.
       return false;
     } else if (utils.isMobile()) {
       return this.shouldShowHoverOnMobile();
@@ -304,13 +310,13 @@ class ItemView extends GalleryComponent {
 
   getItemHover(imageDimensions) {
     const { customHoverRenderer, ...props } = this.props;
-    const shouldHover = this.shouldHover() || null;
+    const shouldHover = this.shouldHover();
     return shouldHover && (
       <ItemHover
         {...props}
         forceShowHover={this.simulateOverlayHover()}
-        shouldHover={shouldHover}
         imageDimensions={imageDimensions}
+        itemWasHovered={this.state.itemWasHovered}
         key="hover"
         actions={{
           handleItemMouseDown: this.handleItemMouseDown,
@@ -590,7 +596,7 @@ class ItemView extends GalleryComponent {
     }
     styles.margin = -styleParams.itemBorderWidth + 'px';
     styles.height = height + 'px';
-    
+
 
     const imageDimensions = this.getImageDimensions();