+++ packages/gallery/tests/components/item/itemHover.spec.js
@@ -1,4 +1,4 @@
-import { utils } from 'pro-gallery-lib';
+import { utils, GALLERY_CONSTS } from 'pro-gallery-lib';
 import { use, expect } from 'chai';
 import spies from 'chai-spies';
 import sinon from 'sinon';
@@ -24,7 +24,6 @@ describe('ItemHover', () => {
     sampleItemViewProps = driver.props.itemView(sampleItem);
     Object.assign(sampleItemViewProps, {
       idx: 1,
-      shouldHover: true,
       imageDimensions: {
         height: `calc(100% - 80px)`,
         marginTop: 10,
@@ -38,18 +37,70 @@ describe('ItemHover', () => {
     });
   });
 
-  it('should rendered or not according to "shouldHover" prop', () => {
+  it('should rendered hover inner only when needed', () => {
     Object.assign(sampleItemViewProps, {
-      shouldHover: false,
+      itemWasHovered: false,
+      renderCustomInfo: undefined,
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      hoveringBehaviour: GALLERY_CONSTS.infoBehaviourOnHover.APPEARS,
+      overlayAnimation: GALLERY_CONSTS.overlayAnimations.SLIDE_UP,
     });
     driver.mount(ItemHover, sampleItemViewProps);
-    expect(driver.find.hook('item-hover-1').length).to.equal(0);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(0);
 
     Object.assign(sampleItemViewProps, {
-      shouldHover: true,
+      itemWasHovered: true,
+      renderCustomInfo: undefined,
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      hoveringBehaviour: GALLERY_CONSTS.infoBehaviourOnHover.APPEARS,
+      overlayAnimation: GALLERY_CONSTS.overlayAnimations.SLIDE_UP,
     });
     driver.mount(ItemHover, sampleItemViewProps);
-    expect(driver.find.hook('item-hover-1').length).to.equal(1);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(0);
+
+    Object.assign(sampleItemViewProps, {
+      itemWasHovered: false,
+      renderCustomInfo: () => {},
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      hoveringBehaviour: GALLERY_CONSTS.infoBehaviourOnHover.APPEARS,
+      overlayAnimation: GALLERY_CONSTS.overlayAnimations.SLIDE_UP,
+    });
+    driver.mount(ItemHover, sampleItemViewProps);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(0);
+
+    Object.assign(sampleItemViewProps, {
+      itemWasHovered: false,
+      renderCustomInfo: () => {},
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      hoveringBehaviour: GALLERY_CONSTS.infoBehaviourOnHover.DISAPPEARS,
+    });
+    driver.mount(ItemHover, sampleItemViewProps);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(1);
+
+    Object.assign(sampleItemViewProps, {
+      itemWasHovered: false,
+      renderCustomInfo: () => {},
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      overlayAnimation: GALLERY_CONSTS.overlayAnimations.NO_EFFECT,
+    });
+    driver.mount(ItemHover, sampleItemViewProps);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(1);
+
+    Object.assign(sampleItemViewProps, {
+      itemWasHovered: true,
+      renderCustomInfo: () => {},
+    });
+    Object.assign(sampleItemViewProps.styleParams, {
+      hoveringBehaviour: GALLERY_CONSTS.infoBehaviourOnHover.APPEARS,
+      overlayAnimation: GALLERY_CONSTS.overlayAnimations.SLIDE_UP,
+    });
+    driver.mount(ItemHover, sampleItemViewProps);
+    expect(driver.find.class('gallery-item-hover-inner').length).to.equal(1);
   });
 
   it('className should be correct', () => {