+++ packages/lib/src/core/presets/bricksGallery.js
@@ -15,7 +15,7 @@ export const fixedStyles = {
   collageDensity: 0.8,
   groupTypes: '1,2h,2v,3t,3b,3l,3r,3v,3h',
 
-  //this params were moved from the presets in layoutHelper and were not tested and checked yet.
+  // this params were moved from the presets in layoutHelper and were not tested and checked yet.
   gallerySize: 400,
   minItemSize: 50,
   chooseBestGroup: true,
@@ -33,6 +33,7 @@ export const createStyles = styles => {
   return {
     ...styles,
     ...fixedStyles,
+    targetItemSize: styles.targetItemSize || styles.gallerySize,
   }
 }