+++ packages/lib/src/core/presets/gridGallery.js
@@ -7,11 +7,11 @@ export const fixedStyles = {
   groupSize: 1,
   hasThumbnails: false,
   groupTypes: '1',
-  //this params were moved from the presets in layoutHelper and were not tested and checked yet.
+  // this params were moved from the presets in layoutHelper and were not tested and checked yet.
   smartCrop: false,
   galleryType: 'Columns',
   fixedColumns: 0,
-  gallerySize: 0,
+  targetItemSize: 0,
   enableScroll: true,
   cropOnlyFill: false,
   isSlider: false,
@@ -26,8 +26,7 @@ export const createStyles = styles => {
   return {
     ...styles,
     ...fixedStyles,
-    gallerySize: styles.modifiedGallerySize ? styles.gallerySize : Math.round(styles.gallerySize * 8.5 + 150),
-    modifiedGallerySize: true
+    targetItemSize: styles.targetItemSize || Math.round(styles.gallerySize * 8.5 + 150),
   }
 }