+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+## [v2.0.43](https://pro-gallery-2-0-43.surge.sh) (03/08/2020)
+ 
+#### GALLERY
+ -  leanGallery: fix isEligible to consider totalItemsCount
+
+---
+## [v2.0.42](https://pro-gallery-2-0-42.surge.sh) (03/08/2020)
+ 
+#### GALLERY
+ -  fix (leanGallery): fix linter error (unknown var)
+ -  improve: rename gallerySize to targetItemSize for inner calculations
+
+---
+## [v2.0.41](https://pro-gallery-2-0-41.surge.sh) (03/08/2020)
+ 
+#### GALLERY
+ -  fixed version 2.0.40: itemHover will be created always (when needed) when there is an overlayAnimation. The inner part of itemHover will be rendered only when needed as well.
+
+---
+## [v2.0.40](https://pro-gallery-2-0-40.surge.sh) (02/08/2020)
+ 
+#### GALLERY
+ -  itemHover will be created always (when needed) when there is an overlayAnimation. The inner part of itemHover will be rendered only on first hover an on.
+
+---
+## [v2.0.39](https://pro-gallery-2-0-39.surge.sh) (02/08/2020)
+ 
+#### OTHER
+ -   better isUsingCustomElement for blueprints
+
+---
+## [v2.0.38](https://pro-gallery-2-0-38.surge.sh) (02/08/2020)
+ 
+#### OTHER
+ -  Created an entry for a non-breaking blueprints-using gallery
+
+#### GALLERY
+ -  leanGallery: update calcItemSize function and added titlePlacement fix for isEligible
+
+---
 ## [v2.0.37](https://pro-gallery-2-0-37.surge.sh) (29/07/2020)
  
 #### GALLERY